move fcnet to kale.embed.nn
