+++ tests/embed/test_nn.py
@@ -1,6 +1,6 @@
 import torch
 
-from kale.embed.nn import RandomLayer
+from kale.embed.nn import FCNet, RandomLayer
 
 
 def test_random_layer_forward():
@@ -20,3 +20,29 @@ def test_random_layer_forward():
     # Check output types and shape
     assert isinstance(output, torch.Tensor), "Output should be a tensor"
     assert output.shape == torch.Size([batch_size, output_dim]), "Output shape should match batch size and output_dim"
+
+
+def test_fcnet_forward():
+    dims = [64, 128, 64]
+    batch_size = 32
+
+    # Initialize FCNet model
+    model = FCNet(dims, dropout=1)
+
+    # Create mock input
+    input_data = torch.randn(batch_size, dims[0])
+
+    # Forward pass through the model
+    output = model(input_data)
+
+    # Check output types and shape
+    assert isinstance(output, torch.Tensor), "Output should be a tensor"
+    assert output.shape == torch.Size([batch_size, dims[-1]]), "Output shape should match batch size and last dim"
+
+
+def test_fcnet_minimal_inputs():
+    # FCNet
+    model = FCNet([1, 1])
+    inp = torch.randn(2, 1)
+    out = model(inp)
+    assert out.shape == (2, 1)
